fix flaky tests on MFA
