+++ e2e/cypress/integration/enterprise/guest_accounts/guest_identification_spec.js
@@ -75,28 +75,11 @@ describe('Guest Accounts', () => {
         cy.findByTestId('ServiceSettings.EnforceMultifactorAuthenticationtrue').check();
 
         // # Click "Save".
-        cy.get('#saveSetting').scrollIntoView().click();
+        cy.findByText('Save').click().wait(TIMEOUTS.ONE_SEC);
 
-        cy.url().then((url) => {
-            if (url.includes('mfa/setup')) {
-                // # Complete MFA setup if we are on token setup page /mfa/setup
-                cy.get('#mfa').wait(TIMEOUTS.HALF_SEC).find('.col-sm-12').then((p) => {
-                    const secretp = p.text();
-                    adminMFASecret = secretp.split(' ')[1];
-
-                    const token = authenticator.generateToken(adminMFASecret);
-                    cy.get('#mfa').find('.form-control').type(token);
-                    cy.get('#mfa').find('.btn.btn-primary').click();
-
-                    cy.wait(TIMEOUTS.HALF_SEC);
-                    cy.get('#mfa').find('.btn.btn-primary').click();
-                });
-            } else {
-                // # If the sysadmin already has MFA enabled, reset the secret.
-                cy.apiGenerateMfaSecret(sysadmin.id).then((res) => {
-                    adminMFASecret = res.code.secret;
-                });
-            }
+        // # Get MFA secret
+        cy.uiGetMFASecret(sysadmin.id).then((secret) => {
+            adminMFASecret = secret;
         });
 
         // # Navigate to Guest Access page.
